Extract generic printing logic from the JavaScript printer
